gitcmd.sh: avoid losing repository
For some reason gitorious was down one day. When my script came along, the
"pull" failed so it deleted the repository and tried recloning, which also
failed, leaving me with no repository. This change will only remove the old
repository if it can clone a new one.

+++ misc/gitcmd.sh
@@ -59,11 +59,16 @@ if [ -n "$FAILED" -a "$GITCMD" = "pull" ]; then
 	echo "FAILED: $FAILED"
 
 	for REP in $FAILED; do
-		echo "deleting and re-cloning fail repository $REP"
+		echo "re-cloning failed repository $REP"
 		pushd $REP
 			REMOTE=$(git remote -v | grep fetch | head -1 | cut -f2 | cut -d' ' -f1)
 		popd
-		rm -fr $REP
-		git clone $REMOTE
+		git clone $REMOTE ${REP}.new
+		if [ $? -eq 0 ]; then
+			rm -fr $REP
+			mv ${REP}.new $REP
+		else
+			echo "clone failure on $REP"
+		fi
 	done
 fi
